requirements.txt
# Core dependencies for audio transcription and news synthesis
requests>=2.31.0
beautifulsoup4>=4.12.0
edge-tts>=6.1.0
pydub>=0.25.0
Pillow>=10.0.0  # For podcast cover image generation

# AI Analysis dependencies for GitHub Actions
anthropic>=0.3.0
aiohttp>=3.8.0

# Optional dependencies for enhanced functionality
# Uncomment if needed for specific features:
# newspaper3k>=0.2.8
# readability-lxml>=0.8.1
# html2text>=2020.1.16
# librosa>=0.10.0  # For advanced audio analysis
# matplotlib>=3.7.0  # For waveform visualization
